+++ NEWS.md
@@ -74,6 +74,9 @@ Standard library changes
 #### Distributed
 
 
+#### UUIDs
+* Change `uuid1` and `uuid4` to use `Random.RandomDevice()` as default random number generator ([#35872]).
+
 Deprecated or removed
 ---------------------